requirements.txt
biopython==1.85
numpy==2.2.6
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
